requirements.txt
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.55.3
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.10.0
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2024.2
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
XlsxWriter==3.2.0
